+++ crates/rome_js_analyze/src/semantic_analyzers/complexity.rs
@@ -1,5 +1,5 @@
 //! Generated file, do not edit by hand, see `xtask/codegen`
 
 use rome_analyze::declare_group;
-mod no_useless_fragments;
+pub(crate) mod no_useless_fragments;
 declare_group! { pub (crate) Complexity { name : "complexity" , rules : [self :: no_useless_fragments :: NoUselessFragments ,] } }