+++ crates/rome_service/src/configuration/generated.rs
@@ -13,9 +13,10 @@ pub(crate) fn push_to_analyzer_rules(
             if let Some(RuleConfiguration::WithOptions(rule_options)) =
                 rules.get_rule_configuration(rule_name)
             {
-                if let Some(options) = &rule_options.options {
+                if let Some(possible_options) = &rule_options.options {
                     if let Some(rule_key) = metadata.find_rule("a11y", rule_name) {
-                        analyzer_rules.push_rule(rule_key, options.to_string());
+                        let rule_options = possible_options.extract_option(&rule_key);
+                        analyzer_rules.push_rule(rule_key, rule_options);
                     }
                 }
             }
@@ -26,9 +27,10 @@ pub(crate) fn push_to_analyzer_rules(
             if let Some(RuleConfiguration::WithOptions(rule_options)) =
                 rules.get_rule_configuration(rule_name)
             {
-                if let Some(options) = &rule_options.options {
+                if let Some(possible_options) = &rule_options.options {
                     if let Some(rule_key) = metadata.find_rule("complexity", rule_name) {
-                        analyzer_rules.push_rule(rule_key, options.to_string());
+                        let rule_options = possible_options.extract_option(&rule_key);
+                        analyzer_rules.push_rule(rule_key, rule_options);
                     }
                 }
             }
@@ -39,9 +41,10 @@ pub(crate) fn push_to_analyzer_rules(
             if let Some(RuleConfiguration::WithOptions(rule_options)) =
                 rules.get_rule_configuration(rule_name)
             {
-                if let Some(options) = &rule_options.options {
+                if let Some(possible_options) = &rule_options.options {
                     if let Some(rule_key) = metadata.find_rule("correctness", rule_name) {
-                        analyzer_rules.push_rule(rule_key, options.to_string());
+                        let rule_options = possible_options.extract_option(&rule_key);
+                        analyzer_rules.push_rule(rule_key, rule_options);
                     }
                 }
             }
@@ -52,9 +55,10 @@ pub(crate) fn push_to_analyzer_rules(
             if let Some(RuleConfiguration::WithOptions(rule_options)) =
                 rules.get_rule_configuration(rule_name)
             {
-                if let Some(options) = &rule_options.options {
+                if let Some(possible_options) = &rule_options.options {
                     if let Some(rule_key) = metadata.find_rule("nursery", rule_name) {
-                        analyzer_rules.push_rule(rule_key, options.to_string());
+                        let rule_options = possible_options.extract_option(&rule_key);
+                        analyzer_rules.push_rule(rule_key, rule_options);
                     }
                 }
             }
@@ -65,9 +69,10 @@ pub(crate) fn push_to_analyzer_rules(
             if let Some(RuleConfiguration::WithOptions(rule_options)) =
                 rules.get_rule_configuration(rule_name)
             {
-                if let Some(options) = &rule_options.options {
+                if let Some(possible_options) = &rule_options.options {
                     if let Some(rule_key) = metadata.find_rule("performance", rule_name) {
-                        analyzer_rules.push_rule(rule_key, options.to_string());
+                        let rule_options = possible_options.extract_option(&rule_key);
+                        analyzer_rules.push_rule(rule_key, rule_options);
                     }
                 }
             }
@@ -78,9 +83,10 @@ pub(crate) fn push_to_analyzer_rules(
             if let Some(RuleConfiguration::WithOptions(rule_options)) =
                 rules.get_rule_configuration(rule_name)
             {
-                if let Some(options) = &rule_options.options {
+                if let Some(possible_options) = &rule_options.options {
                     if let Some(rule_key) = metadata.find_rule("security", rule_name) {
-                        analyzer_rules.push_rule(rule_key, options.to_string());
+                        let rule_options = possible_options.extract_option(&rule_key);
+                        analyzer_rules.push_rule(rule_key, rule_options);
                     }
                 }
             }
@@ -91,9 +97,10 @@ pub(crate) fn push_to_analyzer_rules(
             if let Some(RuleConfiguration::WithOptions(rule_options)) =
                 rules.get_rule_configuration(rule_name)
             {
-                if let Some(options) = &rule_options.options {
+                if let Some(possible_options) = &rule_options.options {
                     if let Some(rule_key) = metadata.find_rule("style", rule_name) {
-                        analyzer_rules.push_rule(rule_key, options.to_string());
+                        let rule_options = possible_options.extract_option(&rule_key);
+                        analyzer_rules.push_rule(rule_key, rule_options);
                     }
                 }
             }
@@ -104,9 +111,10 @@ pub(crate) fn push_to_analyzer_rules(
             if let Some(RuleConfiguration::WithOptions(rule_options)) =
                 rules.get_rule_configuration(rule_name)
             {
-                if let Some(options) = &rule_options.options {
+                if let Some(possible_options) = &rule_options.options {
                     if let Some(rule_key) = metadata.find_rule("suspicious", rule_name) {
-                        analyzer_rules.push_rule(rule_key, options.to_string());
+                        let rule_options = possible_options.extract_option(&rule_key);
+                        analyzer_rules.push_rule(rule_key, rule_options);
                     }
                 }
             }