+++ npm/backend-jsonrpc/src/workspace.ts
@@ -873,7 +873,33 @@ export type RuleConfiguration = RulePlainConfiguration | RuleWithOptions;
 export type RulePlainConfiguration = "warn" | "error" | "off";
 export interface RuleWithOptions {
 	level: RulePlainConfiguration;
-	options: any;
+	options?: PossibleOptions;
+}
+export type PossibleOptions = HooksOptions | null;
+/**
+ * Options for the rule `useExhaustiveDependencies` and `useHookAtTopLevel`
+ */
+export interface HooksOptions {
+	/**
+	 * List of safe hooks
+	 */
+	hooks: Hooks[];
+}
+export interface Hooks {
+	/**
+	* The "position" of the closure function, starting from zero.
+
+### Example 
+	 */
+	closureIndex?: number;
+	/**
+	 * The "position" of the array of dependencies, starting from zero.
+	 */
+	dependenciesIndex?: number;
+	/**
+	 * The name of the hook
+	 */
+	name: string;
 }
 export interface OpenFileParams {
 	content: string;